feat: look for credential helpers in common paths
The docker credential helpers might be present in a few common places
besides the default `/bin:/usr/bin:/usr/sbin` one. Look for it in a few
directories commonly used by package managers.

Looking for a binary if different places is not ideal in terms of
reproductibility but it makes practical sense given the nature of the
executable we're looking for (a credential helper specified in the
user's configuration).