Use platform transition in the oci_image_index rule
The current way of listing images makes Bazel build images for a
platform that is not the configured target platform, defined by
--platforms. This commit changes the behaviour and applies a transition
so that the images are built for the configured target platform.

New way:
oci_image_index(
    name = "app",
    image = ":app_linux",
    platforms = [
        "@io_bazel_rules_go//go/toolchain:linux_amd64",
        "@io_bazel_rules_go//go/toolchain:linux_arm64",
    ]
)

Old way:
oci_image_index(
    name = "app",
    image = [
        ":app_linux_amd64",
        ":app_linux_arm64"
    ],
)

For backwards compatibility, the old way is still possible to use.