Copy files using coreutils cp
Unlike `cat` redirected to a file, `cp` sees both its input and output
and can more readily know the details of its arguments. These details
can sometimes be exploited to achieve zero-copy transfers when the
OS/filesystem can do copy-on-write (e.g. `clonefile` on Mac,
`copy_file_range` on Linux). uutils' `cp` will attempt to take
advantage of these opportunities where possible (see:
https://github.com/uutils/coreutils/blob/35232689366fb102854693b6bd6b6161901d192e/src/uu/cp/src/cp.rs#L2190
).

This is a slight behaviour change with respect to permissions and file
modes: `cp` preserves them while `cat` naturally does not. If file
metadata erasure is an important property of this code (I don't think it
is, but LMK) `cp` could be instructed to use the default file mode/permission
using the `--no-preserve` flag.

Also convert the one of the remaining references to the system's `cat`
(rather than uutils') to also convert it to a call to `cp`.