Support passing headers to download if Bazel supports it
Rather than falling back to curl let's actually pass the headers to the
native Bazel downloader if we're running a version of Bazel that
supports them (i.e. >= 7.1.0). I had to bring the `bazel_features` repo
to detect this.

Note that I've kept support for curl rather than introducing the
breaking change (as #562 discusses) so that we can start taking
advantage of it without breaking for people still on older versions of
Bazel.