Rename oci_pull tags attribute to bazel_tags, remove from oci_alias
On the oci_pull repository rule, the attribute passed in is renamed from
"tags" -> "bazel_tags" to avoid confusion with container/docker tags.

"bazel_tags" is now explicitly defined on the rule as an attribute.

For the oci_alias repository rule, I removed the passed in attribute
since it wasn't being used anywhere.