fix: accumulate image manifests in oci_image_index
In the recent #386, when we replaced use of the "yq" tool with the
"jq" tool, we translated the "oci_image_index" rule's accumulation of
image manifests incorrectly, such that we lose all but the last index
that we add to the index.

In "jq", the "update-assignment" operator (|=) does not append to an
array supplied as its left-hand context value automatically. Instead,
one must use a "complex assignment", mentioning the context value
explicitly together with the addition operator.