Automagically set the created date by annotation
Doing a `docker inspect --format='{{.Created}}' my/image:latest` on an
image built with `oci_image` previously always returned the static
beginning of the unix timestamp (1. January 1970 00:00:00). To be more
precise, that's what it returns for images that are built "from
scratch". When an image is built on a base, then the created date is set
to the created date of the base image.

Both are fine strategies for reproducible builds, but in stamped builds
that are eventually shipped/deployed, you probably want to set that time
to the actual build time.

Stamping is already supported, e.g. by setting annotations, which can
use stamp variables through a workspace status command.

Furthermore, the OCI image spec already [defines an annotation](https://github.com/opencontainers/image-spec/blob/main/annotations.md#pre-defined-annotation-keys) to describe the time and date the image was built.
Which of course, when `docker load -i mytarball.tar` is dutifully
ignored by docker.

We can leverage this and, if the OCI annotation is provided, also set
the created date to that same value. When it isn't, we simply fall back
to the old behaviour.