Introduce new attribute for `created` datetime
Doing a docker inspect --format='{{.Created}}' my/image:latest on an
image built with oci_image previously always returned the static
beginning of the unix timestamp (1. January 1970 00:00:00). To be more
precise, that's what it returns for images that are built "from
scratch". When an image is built on a base, then the created date is set
to the created date of the base image.

Both are fine strategies for reproducible builds, but in stamped builds
that are eventually shipped/deployed, you probably want to set that time
to the actual build time.

To allow setting that, we introduce a separate file label, allowing to
pass in a file containing a stamp variable.