Suppress regctl stderr in image.sh (#863)
If `regctl` is unable to access the default Docker credentials file it
prints a warning message about this to stderr on every invocation
despite none of the usages in `image.sh` requiring credentials. Redirect
stderr to `/dev/null` for all `regctl` invocations in this script to
avoid build log spam in this situation.